Reorganize Resolvers
Resolvers and Queries were reorganized. Now each Ecto schema has a file for queries and mutations and another for resolvers.

closes #48
